The Shoes Spot
My classmate asked me to make a simple website for our school task, which is our final project, and he offered to pay me a bit. I agreed and created a website for him for our final project at our school. It turned out well, and he was happy with it. I enjoyed working on the project and learning more about making websites.